fix: reuse shared ScanBreakdown controls and harden scanner/Rust JSON contracts
WinUI: replace duplicated Largest-Files/Top-Folders/breakdown XAML and
open/filter/sort handlers in ScanPage and HistoryPage with the shared
LargestFilesControl, TopFoldersControl, and BreakdownListControl.

.NET: change StreamComplete.TotalFiles int->long (parity with
ScanResult.TotalFiles) so a >int.MaxValue streaming count no longer
overflows during parse and silently drops the final result; add
ScanHistoryRecord contract tests and a total_files overflow regression.

Rust scanner: honor --top end-to-end via ScanOptions.top_n, exclude
Windows hidden-attribute files (not just dotfiles), run --min-size/
--include-hidden before the cache hit, match categories by whole path
component, and populate category_sizes on the CPU path; remove the dead,
broken scan_directory_parallel. Verified: cargo test -p shared-scanner
--features gpu passed (incl. new integration tests).

+++ gui-winui/SpaceAnalyzer/Models/StreamEvent.cs
@@ -69,8 +69,12 @@ public class StreamComplete
     [JsonPropertyName("type")]
     public string Type { get; set; } = "complete";
 
+    // Long (not int): parallels ScanResult.TotalFiles. A streaming scan whose
+    // complete event reports more than int.MaxValue files would overflow an int
+    // during JSON parsing, which the streaming reader swallows — silently
+    // dropping the final result. Same class of bug as the file_types overflow.
     [JsonPropertyName("total_files")]
-    public int TotalFiles { get; set; }
+    public long TotalFiles { get; set; }
 
     [JsonPropertyName("total_size_bytes")]
     public ulong TotalSizeBytes { get; set; }
